updated version

+++ new_release.sh
@@ -1,9 +1,15 @@
 # update the next version in the variables file (variable CURRENT_VERSION)
 
-rm build/bin/*
+echo -n "Did you update the version in the variables file ? " 
+read updated_version
 
-wails build -platform "windows/amd64" -nsis -o "hatt-windows-amd64.exe"
+if [ $updated_version == 'y' ]; then
 
-mv ./build/bin/hatt-amd64-installer.exe ./build/bin/hatt-windows-amd64-installer.exe
+    rm build/bin/*
 
-wails build -platform "linux/amd64" -o "hatt-linux-amd64"
+    wails build -platform "windows/amd64" -nsis -o "hatt-windows-amd64.exe"
+
+    mv ./build/bin/hatt-amd64-installer.exe ./build/bin/hatt-windows-amd64-installer.exe
+
+    wails build -platform "linux/amd64" -o "hatt-linux-amd64"
+fi
